backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pymongo==4.6.0
motor==3.3.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
slowapi==0.1.8
sqlalchemy==2.0.23
email-validator==2.1.0.post1
python-multipart==0.0.6
emails==0.6 